% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Title ORs & 95% CIs}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{beta coefficient}

\item{se}{standard error}

\item{siglevel}{significance level}

\item{roundto}{round to decimal places}
}
\value{
ORresult
}
\description{
Title ORs & 95% CIs
}
\examples{
OR_95CI(m1coefs, m1ses, siglevel = 0.05, roundto = 2)

}
\author{
Syeda Aiman Fatima
}
